+++ rasa_nlu/evaluate.py
@@ -20,11 +20,15 @@ logger = logging.getLogger(__name__)
 
 
 def create_argparser():
-    parser = argparse.ArgumentParser(description='evaluate a trained Rasa NLU pipeline')
+    parser = argparse.ArgumentParser(
+            description='evaluate a trained Rasa NLU pipeline')
 
-    parser.add_argument('-d', '--data', default=None, help="file containing evaluation data")
-    parser.add_argument('-m', '--model', required=True, help="path to model")
-    parser.add_argument('-c', '--config', required=True, help="config file")
+    parser.add_argument('-d', '--data', default=None,
+                        help="file containing evaluation data")
+    parser.add_argument('-m', '--model', required=True,
+                        help="path to model")
+    parser.add_argument('-c', '--config', required=True,
+                        help="config file")
     return parser
 
 
@@ -33,14 +37,15 @@ def plot_confusion_matrix(cm, classes,
                           title='Confusion matrix',
                           cmap=None,
                           zmin=1):
-    """This function prints and plots the confusion matrix for the intent classification.
+    """Print and plot the confusion matrix for the intent classification.
 
     Normalization can be applied by setting `normalize=True`."""
     import numpy as np
 
     zmax = cm.max()
-    plt.imshow(cm, interpolation='nearest', cmap=cmap if cmap else plt.cm.Blues, aspect='auto',
-               norm=LogNorm(vmin=zmin, vmax=zmax))
+    plt.clf()
+    plt.imshow(cm, interpolation='nearest', cmap=cmap if cmap else plt.cm.Blues,
+               aspect='auto', norm=LogNorm(vmin=zmin, vmax=zmax))
     plt.title(title)
     plt.colorbar()
     tick_marks = np.arange(len(classes))
@@ -64,16 +69,18 @@ def plot_confusion_matrix(cm, classes,
 
 
 def log_evaluation_table(test_y, preds):
-    from sklearn.metrics import accuracy_score
-    from sklearn.metrics import classification_report
-    from sklearn.metrics import f1_score
-    from sklearn.metrics import precision_score
+    from sklearn import metrics
+
+    report = metrics.classification_report(test_y, preds)
+    precision = metrics.precision_score(test_y, preds, average='weighted')
+    f1 = metrics.f1_score(test_y, preds, average='weighted')
+    accuracy = metrics.accuracy_score(test_y, preds)
 
     logger.info("Intent Evaluation Results")
-    logger.info("F1-Score:  {}".format(f1_score(test_y, preds, average='weighted')))
-    logger.info("Precision: {}".format(precision_score(test_y, preds, average='weighted')))
-    logger.info("Accuracy:  {}".format(accuracy_score(test_y, preds)))
-    logger.info("Classification report: \n{}".format(classification_report(test_y, preds)))
+    logger.info("F1-Score:  {}".format(f1))
+    logger.info("Precision: {}".format(precision))
+    logger.info("Accuracy:  {}".format(accuracy))
+    logger.info("Classification report: \n{}".format(report))
 
 
 def run_intent_evaluation(config, model_path, component_builder=None):
@@ -81,7 +88,7 @@ def run_intent_evaluation(config, model_path, component_builder=None):
     from sklearn.utils.multiclass import unique_labels
 
     # get the metadata config from the package data
-    test_data = load_data(config['data'])
+    test_data = load_data(config['data'], config['language'])
     interpreter = Interpreter.load(model_path, config, component_builder)
 
     test_y = [e.get("intent") for e in test_data.training_examples]