+++ tests/base/test_featurizers.py
@@ -16,10 +16,9 @@ from rasa_nlu.training_data import Message
     ("hey how are you today", [-0.19649599, 0.32493639, -0.37408298, -0.10622784, 0.062756])
 ])
 def test_spacy_featurizer(sentence, expected, spacy_nlp):
-    from rasa_nlu.featurizers.spacy_featurizer import SpacyFeaturizer
-    ftr = SpacyFeaturizer()
+    from rasa_nlu.featurizers import spacy_featurizer
     doc = spacy_nlp(sentence)
-    vecs = ftr.features_for_doc(doc)
+    vecs = spacy_featurizer.features_for_doc(doc)
     assert np.allclose(doc.vector[:5], expected, atol=1e-5)
     assert np.allclose(vecs, doc.vector, atol=1e-5)